CI: Increase timeout for executing STM32 unit test
It seems that with newer versions of the STM32-related tools, especially
openocd, it takes much longer to transfer the debug output from uD3TN to
the debugger's stdout. Thus, we have to increase the timeout to keep the
unittests running until their success (or failure) can be properly
indicated to the debugger.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>