Add Bundle Age Extension Block
The Bundle Age block [...] contains the number of milliseconds
that have elapsed between the time the bundle was created and time at
which it was most recently forwarded. It is intended for use by nodes
lacking access to an accurate clock, to aid in determining the time at
which a bundle's lifetime expires. [...] If the bundle's creation time
is zero, then the bundle MUST contain exactly one (1) occurrence of this
type of block.[1]

In this implementation the age block is only evaluated if the creation
timestamp is 0. If a bundle age block exists, it will be updated
according to the specification, but an age block will never be
proactively inserted.

[1]: https://tools.ietf.org/html/draft-ietf-dtn-bpbis-30#section-4.4.2

Signed-off-by: Maximilian Nitsch <maximilian.nitsch@d3tn.com>